separate runtime config
