+++ runtime/selendra/constants/src/lib.rs
@@ -24,6 +24,8 @@ pub mod weights;
 pub mod currency {
 	use primitives::Balance;
 
+	pub const EXISTENTIAL_DEPOSIT: Balance = 1 * DOLLARS;
+
 	pub const MILLICENTS: Balance = 1_000_000_000;
 	pub const CENTS: Balance = 1_000 * MILLICENTS; // assume this is worth about a cent.
 	pub const DOLLARS: Balance = 100 * CENTS;
